requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
boto3==1.26.114
botocore==1.29.114
certifi==2022.12.7
coverage==7.2.5
charset-normalizer==3.1.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.3.0
Django==4.1.7
django-filter==23.1
django-import-export==3.1.0
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
ipython==8.12.0
jedi==0.18.2
jmespath==1.0.1
MarkupPy==1.14
matplotlib-inline==0.1.6
multidict==6.0.4
odfpy==1.4.1
openpyxl==3.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.4.0
traitlets==5.9.0
twilio==8.0.0
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
